pyproject.toml
[tool.poetry]
name = "emulator"
version = "0.1.0"
description = ""
authors = ["Ben Mz <bmz@prohacktive.io>"]

[tool.poetry.dependencies]
python = "^3.6"
pyzmq = "^19.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
